chore: prepare v2 upgrade
