--- app/upgrades/v1_5/v1_4_types/funders/getters_params.go
@@ -1,28 +1,0 @@
-package funders
-
-import (
-	storeTypes "cosmossdk.io/store/types"
-	"github.com/KYVENetwork/chain/x/funders/types"
-	"github.com/cosmos/cosmos-sdk/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-)
-
-// GetParams get all parameters as types.Params
-func GetParams(ctx sdk.Context, cdc codec.Codec, storeKey storeTypes.StoreKey) (params Params) {
-	store := ctx.KVStore(storeKey)
-
-	bz := store.Get(types.ParamsKey)
-	if bz == nil {
-		return params
-	}
-
-	cdc.MustUnmarshal(bz, &params)
-	return params
-}
-
-// SetParams set the params
-func SetParams(ctx sdk.Context, cdc codec.Codec, storeKey storeTypes.StoreKey, params Params) {
-	store := ctx.KVStore(storeKey)
-	bz := cdc.MustMarshal(&params)
-	store.Set(types.ParamsKey, bz)
-}
