+++ app/upgrades/v2_0/upgrade.go
@@ -1,0 +1,36 @@
+package v2_0
+
+import (
+	"context"
+	"fmt"
+
+	"cosmossdk.io/log"
+	upgradetypes "cosmossdk.io/x/upgrade/types"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/module"
+)
+
+const (
+	UpgradeName = "v1.6.0"
+)
+
+var logger log.Logger
+
+func CreateUpgradeHandler(
+	mm *module.Manager,
+	configurator module.Configurator,
+) upgradetypes.UpgradeHandler {
+	return func(ctx context.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
+		sdkCtx := sdk.UnwrapSDKContext(ctx)
+		logger = sdkCtx.Logger().With("upgrade", UpgradeName)
+		logger.Info(fmt.Sprintf("performing upgrade %v", UpgradeName))
+
+		// Run cosmos migrations
+		migratedVersionMap, err := mm.RunMigrations(ctx, configurator, fromVM)
+
+		// Run KYVE migrations
+
+		return migratedVersionMap, err
+	}
+}
