+++ interchaintest/upgrades/v1_5/upgrade_utils_test.go
@@ -1,4 +1,4 @@
-package v1_5_test
+package v2_0_test
 
 import (
 	"encoding/json"
